+++ wp-content/plugins/contact-form-7-mailchimp-extension/cf7-mch-ext.php
@@ -7,7 +7,7 @@ Author: Renzo Johnson
 Author URI: http://renzojohnson.com
 Text Domain: contact-form-7
 Domain Path: /languages/
-Version: 0.4.37
+Version: 0.4.38
 */
 
 /*  Copyright 2013-2017 Renzo Johnson (email: renzojohnson at gmail.com)
@@ -27,7 +27,7 @@ Version: 0.4.37
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-define( 'SPARTAN_MCE_VERSION', '0.4.37' );
+define( 'SPARTAN_MCE_VERSION', '0.4.38' );
 define( 'SPARTAN_MCE_PLUGIN_BASENAME', plugin_basename( __FILE__ ) );
 define( 'SPARTAN_MCE_PLUGIN_NAME', trim( dirname( SPARTAN_MCE_PLUGIN_BASENAME ), '/' ) );
 define( 'SPARTAN_MCE_PLUGIN_DIR', untrailingslashit( dirname( __FILE__ ) ) );
@@ -38,8 +38,8 @@ require_once( SPARTAN_MCE_PLUGIN_DIR . '/lib/mailchimp.php' );
 
 function mc_meta_links( $links, $file ) {
     if ( $file === 'contact-form-7-mailchimp-extension/cf7-mch-ext.php' ) {
-        $links[] = '<a href="'.MCE_URL.'" target="_blank" title="Documentation">Documentation</a>';
-        // $links[] = '<a href="'.MCE_URL.'" target="_blank" title="Starter Guide">Starter Guide</a>';
+        // $links[] = '<a href="'.MCE_URL.'" target="_blank" title="Documentation">Documentation</a>';
+        $links[] = '<a href="'.MCE_URL.'" target="_blank" title="Starter Guide">Starter Guide</a>';
         $links[] = '<a href="//www.paypal.me/renzojohnson" target="_blank" title="Donate"><strong>Donate</strong></a>';
     }
     return $links;
@@ -59,12 +59,4 @@ function mc_after_setup_theme() {
      add_filter('plugin_action_links_' . SPARTAN_MCE_PLUGIN_BASENAME, 'mc_settings_link');
 }
 add_action ('after_setup_theme', 'mc_after_setup_theme');
-
-
-
-
-
-
-
-