+++ wp-content/plugins/updraftplus/includes/updraftplus-notices.php
@@ -276,6 +276,7 @@ class UpdraftPlus_Notices extends Updraft_Notices {
 	
 	protected function wp_optimize_installed($plugin_base_dir = null, $product_name = null) {
 		$wp_optimize_file = false;
+		if (!function_exists('get_plugins')) require_once(ABSPATH.'wp-admin/includes/plugin.php');
 		$plugins = get_plugins();
 
 		foreach ($plugins as $key => $value) {