+++ wp-content/plugins/updraftplus/options.php
@@ -24,12 +24,13 @@ class UpdraftPlus_Options {
 	}
 
 	public static function get_updraft_option($option, $default = null) {
-		return get_option($option, $default);
+		$ret = get_option($option, $default);
+		return apply_filters('updraftplus_get_option', $ret, $option, $default);
 	}
 
 	// The apparently unused parameter is used in the alternative class in the Multisite add-on
 	public static function update_updraft_option($option, $value, $use_cache = true) {
-		return update_option($option, $value);
+		return update_option($option, apply_filters('updraftplus_update_option', $value, $option, $use_cache));
 	}
 
 	public static function delete_updraft_option($option) {